Code refactoring around render auto-calc fields; and autosave.
