+++ cccm-frontend/src/components/form/formSections/FormNavigation.vue
@@ -53,9 +53,7 @@ export default {
     parentNavJumpToPointed: {
       type: String,
       default: '0',
-    },
-    continueEmitParentNavClicked: true,
-    pageRefreshSectionIndex: ''
+    }
   },
   data() {
     return {
@@ -182,14 +180,9 @@ export default {
       this.currentSectionChild = posChildNav;
 
       // emit an event, parentNavClicked, to the parent, so parent knows the currentSectionParent
-      //console.log('parentNavClicked: ', this.currentSectionParent, this.continueEmitParentNavClicked, this.pageRefreshSectionIndex);
-      // if pageRefreshSectionIndex is not set, or currentSection != pageRefreshSectionIndex; then emit the event
-      if (this.pageRefreshSectionIndex == '' || this.currentSectionParent != this.pageRefreshSectionIndex) {
-        this.$emit('parentNavClicked', this.currentSectionParent);
-      } else if (this.currentSectionParent == this.pageRefreshSectionIndex && this.continueEmitParentNavClicked) {
-        this.$emit('refreshPage');
-      } 
-                          
+      //console.log('parentNavClicked: ', this.currentSectionParent);
+      this.$emit('parentNavClicked', this.currentSectionParent);
+                                
       // show/hide sideCards panels
       //this.showHideRightsidePanels();
